+++ src/main/kotlin/icu/windea/ut/toolbox/jast/JsonPointerBasedReferenceProvider.kt
@@ -1,11 +1,14 @@
 package icu.windea.ut.toolbox.jast
 
-import com.intellij.codeInsight.highlighting.HighlightedReference
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.*
+import com.intellij.psi.impl.source.resolve.ResolveCache
 import com.intellij.util.IncorrectOperationException
 import com.intellij.util.ProcessingContext
 
+/**
+ * @see JsonPointerBasedLanguageSettings.references
+ */
 class JsonPointerBasedReferenceProvider : PsiReferenceProvider() {
     override fun getReferencesByElement(element: PsiElement, context: ProcessingContext): Array<PsiReference> {
         val jElement = element.toJElement()
@@ -23,11 +26,9 @@ class JsonPointerBasedReferenceProvider : PsiReferenceProvider() {
             else -> jElement.psi.text
         }
         if(text.isNullOrEmpty()) return PsiReference.EMPTY_ARRAY
-        val name = JsonPointerManager.getNameForLanguageSettings(jElement)
+        val (name, textOffset) = jElement.getNameAndTextOffset()
         if(name.isNullOrEmpty()) return PsiReference.EMPTY_ARRAY
-        val nameInTextOffset = text.indexOf(name)
-        if(nameInTextOffset == -1) return PsiReference.EMPTY_ARRAY
-        val range = TextRange.from(startOffset + nameInTextOffset, name.length)
+        val range = TextRange.from(startOffset + textOffset, name.length)
         val currentFile = element.containingFile ?: return PsiReference.EMPTY_ARRAY
 
         return arrayOf(Reference(element, range, currentFile, jElement, name, languageSettings))
@@ -36,20 +37,34 @@ class JsonPointerBasedReferenceProvider : PsiReferenceProvider() {
     class Reference(
         element: PsiElement,
         range: TextRange,
-        private val currentFile: PsiFile,
-        private val jElement: JElement,
-        private val name: String,
-        private val languageSettings: JsonPointerBasedLanguageSettings
-    ) : PsiPolyVariantReferenceBase<PsiElement>(element, range), HighlightedReference {
+        val currentFile: PsiFile,
+        val jElement: JElement,
+        val name: String,
+        val languageSettings: JsonPointerBasedLanguageSettings
+    ) : PsiPolyVariantReferenceBase<PsiElement>(element, range) {
+        val project by lazy { element.project }
+        
         override fun handleElementRename(newElementName: String): PsiElement {
             throw IncorrectOperationException()
         }
+        
+        //cached
 
-        override fun multiResolve(incompleteCode: Boolean): Array<ResolveResult> {
+        private object MultiResolver : ResolveCache.PolyVariantResolver<Reference> {
+            override fun resolve(ref: Reference, incompleteCode: Boolean): Array<out ResolveResult> {
+                return ref.doMultiResolve()
+            }
+        }
+
+        override fun multiResolve(incompleteCode: Boolean): Array<out ResolveResult> {
+            return ResolveCache.getInstance(project).resolveWithCaching(this, MultiResolver, false, false)
+        }
+
+        private fun doMultiResolve(): Array<out ResolveResult> {
             val result = mutableSetOf<ResolveResult>()
             languageSettings.references.forEach {
                 JsonPointerManager.processElements(it, currentFile) { resolved ->
-                    val resolvedName = JsonPointerManager.getNameForLanguageSettings(resolved)
+                    val (resolvedName) = resolved.getNameAndTextOffset()
                     if(name == resolvedName) {
                         result += PsiElementResolveResult(resolved.psi)
                     }