+++ src/main/kotlin/icu/windea/ut/toolbox/jast/_extensions.kt
@@ -3,12 +3,24 @@
 package icu.windea.ut.toolbox.jast
 
 import com.intellij.psi.PsiElement
+import icu.windea.ut.toolbox.util.Tuple2
 
 inline fun PsiElement.toJElement(): JElement? = JElementManager.getJElement(this)
 
 inline fun <reified T : JElement> PsiElement.toJElementOfType(): T? = JElementManager.getJElement(this, T::class.java)
 
 inline fun <T : JElement> PsiElement.toJElementOfTypes(vararg targetTypes: Class<out T>): JElement? = JElementManager.getJElement(this, *targetTypes)
+
+//JAST Extensions
+
+fun JElement.getNameAndTextOffset(): Tuple2<String?, Int> {
+    return when {
+        this is JProperty -> this.keyElement?.let { it.value to it.textOffset } ?: (null to 0)
+        this is JPropertyKey -> this.let { it.value to it.textOffset }
+        this is JString -> this.let { it.value to it.textOffset }
+        else -> null to 0
+    }
+}
 
 //JAST Search Extensions