+++ src/main/kotlin/icu/windea/ut/toolbox/jsonSchema/JsonSchemaJsonPointerBasedLanguageSettingsProvider.kt
@@ -23,6 +23,9 @@ class JsonSchemaJsonPointerBasedLanguageSettingsProvider : JsonPointerBasedLangu
         val node = schemaNode.get("\$languageSettings") ?: return null
         return JsonPointerBasedLanguageSettings(
             references = node.get("references")?.toStringOrStringSetValue().orEmpty(),
+            hintForReferences = node.get("hintForReferences")?.booleanValue() ?: true,
+            inspectionForReferences = node.get("inspectionForReferences")?.booleanValue() ?: true,
+            completionForReferences = node.get("completionForReferences")?.booleanValue() ?: true,
         )
     }