+++ CHANGELOG.md
@@ -7,8 +7,10 @@
 * [X] 实现JAST，方面后续对类似JSON的文件格式统一提供扩展
   * 参见：`icu.windea.ut.toolbox.jast.JElementProvider`
   * 默认支持的文件格式：JSON、YAML
-* [X] 扩展jsonSchema，提供基于ANT路径表达式与JSON指针的引用解析
+* [X] 扩展jsonSchema，提供（基于JSON指针的）语言设置，方面后续对类似JSON的文件格式统一提供扩展
   * 参见：`icu.windea.ut.toolbox.jast.JsonPointerBasedLanguageSettingsProvider` 
-  * 参见：`icu.windea.ut.toolbox.jsonSchema.JsonSchemaJsonPointerBasedLanguageSettingsProvider` 
+  * 参见：`icu.windea.ut.toolbox.jsonSchema.JsonSchemaJsonPointerBasedLanguageSettingsProvider`
+* [X] 提供基于（基于JSON指针的）语言设置的引用解析，以及适用于这些引用的语言高亮与代码检查
   * 参见：`icu.windea.ut.toolbox.jast.JsonPointerBasedReferenceProvider`
-  * 默认支持的文件格式：JSON、YAML
+  * 参见：`icu.windea.ut.toolbox.jast.JsonPointerBasedReferenceAnnotator`
+  * 参见：`icu.windea.ut.toolbox.jast.UnresolvedJsonPointerBasedReferenceInspection`